.history/requirements_20240527041652.txt
selenium
webdriver_manager
pandas
sqlalchemy
psycopg2
langdetect



#backend requirements 
fastapi
python-dotenv

apache-airflow
faust
